More squid experiments/fixes (#448)
* Fixed a few small issues:
fixed capitalization so it recognizes that it can export AnimSequences
made it not throw an error when you try to import a psk with a SkeletalMesh that has no sockets selected.
Made it so it can (usually) find the material you want when importing a psk if the material is an import, rather than an export.

* added a few more experiments. not sure if they work fully yet.

* fixed a few bugs in psk import. you should no longer see crinkly meshes.

* committing the rest of the mostly working and widely applicable stuff I have floating around.

* fixed one small issue that blocked build

* removed a bit that got added back in somehow.